Added Tray Icon, Added notes about using RAMDisk
Added a tray icon (if it exists).
Started using RAMDisk to prevent using harddrive for the ping results to output to. This would have resulted in many unnecessary read/writes to the harddrive.
Added some troubleshooting lines that are commented out and tagged "troubleshooting".
Updated functionality so that the pingfile is now deleted every 500 lines so that it won't get too big. This functionality includes the necessary killing of the ping process that's running in the background.